set openai.proxy explicitly to Requests lib to override environmental defaults
By default, the Requests library use the proxy settings from the environment variable HTTPS_PROXY.
It has higher priority than the "proxies" setting in the session, which may lead to confusing behaviors
if both environment variables and the "openai.proxy" config are set.
This patch explicitly informs the Requests library to use the "openai.proxy" config.