Periodically refresh open `requests.Session`s to mitigate open filehandle issues (#179)
As reported from the field - we create a `requests.Session` object on first request to the servers and then reuse it indefinitely. This can leave some open
file handles on the OS (not a big deal), but can interact poorly with a bug in Docker Desktop which causes the SDK to entierly break connections to the server.
See https://github.com/openai/openai-python/issues/140 for more info.