Fix test_file_cli on Windows
On Windows (only), a file created by NamedTemporaryFile can't be
opened again while it is already open. That causes some tests to
fail with PermissionError even when the code under test is correct.

To fix it without making the test code longer or more complicated,
we can use the tmp_path pytest fixture, which makes a temporary
directory (separate for each test and each run) where we can create
whatever file(s) we need, and which handles cleanup. This has the
further benefit of being a more idiomatic way to work with
temporary files in pytest tests.

This commit makes that change for test_file_cli.