Fix api_key_path tests on Windows
As in test_file_cli, these failed with PermissionError on Windows
by trying to open a file created by NamedTemporary file while it is
already open. This likewise switches to using the pytest tmp_path
fixture. The api_key_file fixture is modified to request and use
it, and to provide a Path object, which test case functions call
write_text on instead of printing and flushing.

Since this was a major change to the api_key_file fixture, I've
also taken the opportunity to have it delegate patching and
automatically unpatching openai.api_key_path to another pytest
fixture, allowing its code to be shorter and simpler.