Fix command syntax in test_long_examples_validator
This fixes one of the problems running test_long_examples_validator
on Windows: the command run by subprocess.run giving a "The
filename, directory name, or volume label syntax is incorrect"
error.

It happens because the command to run is specified as a list of one
element that is a string for the shell to interpret, rather than
that string itself. This ends up working on Unix-like systems, but
on Windows it results in extra quotation (adjacent `"` characters)
that cmd.exe doesn't understand.

This changes the subprocess.run call to what was probably intended,
but it may make sense to modify it further to make it more
consistent with how the subprocess module is used in other tests.

Note that test_long_examples_validator is not fully fixed; now it
raises PermissionError on Windows (the NamedTemporaryFile issue).