Fix UnicodeDecodeError from subprocess in Windows
In test_long_examples_validator, subprocess.run was treating the
child process's standard streams as UTF-8 encoded. It's usually
good practice to specify UTF-8, but in this case, the child
process uses a different encoding for its standard streams on
Windows.

On many Windows systems, this is CP-1252 (Windows-1252), and the
curly apostrophe in the "After you’ve fine-tuned a model..."
message triggers a UnicodeDecodeError. Although the message could
be changed, the test's correcteness probably shouldn't rely on
knowledge of the output besides what we specifically test.

This has subprocess.run use the system default encoding for the
child process's standard streams. Since the child process is itself
Python, this should match in nearly all circumstances.