Fix subprocess TypeError (Windows, Python 3.7)
When using Python 3.7 on Windows (but not on other systems, and not
on Windows with any later version of Python), subprocess.run and
related functions don't support Path objects as command line
arguments, due to https://github.com/python/cpython/issues/85919.

This explicitly converts them to str in those calls. This change
can be undone whenever this codebase drops Python 3.7 support.